Refactor move cors and timeout
