+++ zio-http/src/main/scala/zhttp/http/Middleware.scala
@@ -1,5 +1,7 @@
 package zhttp.http
 
+import io.netty.handler.codec.http.HttpHeaderNames
+import zhttp.http.CORS.DefaultCORSConfig
 import zio.clock.Clock
 import zio.duration.Duration
 import zio.{UIO, ZIO}
@@ -82,8 +84,6 @@ object Middleware {
 
   def fromHttp[R, E, A, B](http: Http[R, E, A, B]): Middleware[R, E, Nothing, Any, A, B] = Constant(http)
 
-  def identity: Middleware[Any, Nothing, Nothing, Any, Any, Nothing] = Identity
-
   def ifThenElse[A]: PartialIfThenElse[A] = new PartialIfThenElse(())
 
   def ifThenElseZIO[A]: PartialIfThenElseZIO[A] = new PartialIfThenElseZIO(())
@@ -97,6 +97,91 @@ object Middleware {
   def makeZIO[A]: PartialMakeZIO[A] = new PartialMakeZIO[A](())
 
   def succeed[B](b: B): Middleware[Any, Nothing, Nothing, Any, Any, B] = fromHttp(Http.succeed(b))
+
+  /**
+   * Creates a new constants middleware that always executes the app provided, independent of where the middleware is
+   * applied
+   */
+  def fromApp[R, E](app: HttpApp[R, E]): HttpMiddleware[R, E] = Middleware.fromHttp(app)
+
+  /**
+   * Times out the application with a 408 status code.
+   */
+  def timeout(duration: Duration): HttpMiddleware[Clock, Nothing] =
+    Middleware.identity.race(Middleware.fromApp(Http.status(Status.REQUEST_TIMEOUT).delayAfter(duration)))
+
+  /**
+   * Creates a new middleware using a function from request parameters to a HttpMiddleware
+   */
+  def fromMiddlewareFunction[R, E](f: RequestP[HttpMiddleware[R, E]]): HttpMiddleware[R, E] =
+    Middleware.fromMiddlewareFunctionZIO((method, url, headers) => UIO(f(method, url, headers)))
+
+  /**
+   * Creates a new middleware using a function from request parameters to a ZIO of HttpMiddleware
+   */
+  def fromMiddlewareFunctionZIO[R, E](f: RequestP[ZIO[R, E, HttpMiddleware[R, E]]]): HttpMiddleware[R, E] =
+    Middleware.makeZIO[Request](req => f(req.method, req.url, req.getHeaders))
+
+  /**
+   * An empty middleware that doesn't do anything
+   */
+  def identity: Middleware[Any, Nothing, Nothing, Any, Any, Nothing] = Middleware.Identity
+
+  /**
+   * Creates a middleware for Cross-Origin Resource Sharing (CORS).
+   * @see
+   *   https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS
+   */
+  def cors[R, E](config: CORSConfig = DefaultCORSConfig): HttpMiddleware[R, E] = {
+    def allowCORS(origin: Header, acrm: Method): Boolean                           =
+      (config.anyOrigin, config.anyMethod, origin._2.toString, acrm) match {
+        case (true, true, _, _)           => true
+        case (true, false, _, acrm)       =>
+          config.allowedMethods.exists(_.contains(acrm))
+        case (false, true, origin, _)     => config.allowedOrigins(origin)
+        case (false, false, origin, acrm) =>
+          config.allowedMethods.exists(_.contains(acrm)) &&
+          config.allowedOrigins(origin)
+      }
+    def corsHeaders(origin: Header, method: Method, isPreflight: Boolean): Headers = {
+      Headers.ifThenElse(isPreflight)(
+        onTrue = config.allowedHeaders.fold(Headers.empty) { h =>
+          Headers(HttpHeaderNames.ACCESS_CONTROL_ALLOW_HEADERS.toString(), h.mkString(","))
+        },
+        onFalse = config.exposedHeaders.fold(Headers.empty) { h =>
+          Headers(HttpHeaderNames.ACCESS_CONTROL_EXPOSE_HEADERS.toString(), h.mkString(","))
+        },
+      ) ++
+        Headers(HttpHeaderNames.ACCESS_CONTROL_ALLOW_ORIGIN.toString(), origin._2) ++
+        Headers(
+          HttpHeaderNames.ACCESS_CONTROL_ALLOW_METHODS.toString(),
+          config.allowedMethods.fold(method.toString())(m => m.map(m => m.toString()).mkString(",")),
+        ) ++
+        Headers.when(config.allowCredentials) {
+          Headers(HttpHeaderNames.ACCESS_CONTROL_ALLOW_CREDENTIALS, config.allowCredentials.toString)
+        }
+    }
+    Middleware.fromMiddlewareFunction((method, _, headers) => {
+      (
+        method,
+        headers.getHeader(HttpHeaderNames.ORIGIN),
+        headers.getHeader(HttpHeaderNames.ACCESS_CONTROL_REQUEST_METHOD),
+      ) match {
+        case (Method.OPTIONS, Some(origin), Some(acrm)) if allowCORS(origin, Method.fromString(acrm._2.toString)) =>
+          fromApp(
+            Http.succeed(
+              Response(
+                Status.NO_CONTENT,
+                headers = corsHeaders(origin, Method.fromString(acrm._2.toString), isPreflight = true),
+              ),
+            ),
+          )
+        case (_, Some(origin), _) if allowCORS(origin, method)                                                    =>
+          HttpMiddleware.addHeader(corsHeaders(origin, method, isPreflight = false))
+        case _ => Middleware.identity
+      }
+    })
+  }
 
   private[zhttp] def execute[R, E, AIn, BIn, AOut, BOut](
     http: Http[R, E, AIn, BIn],
